+++ src/models.py
@@ -1,0 +1,41 @@
+
+from sqlalchemy import Column, Integer, String, Float, DateTime, ForeignKey, Table, func
+from sqlalchemy.orm import relationship
+from database import Base
+
+user_cities = Table(
+    "user_cities",
+    Base.metadata,
+    Column("user_id", Integer, ForeignKey("users.id"), primary_key=True),
+    Column("city_id", Integer, ForeignKey("cities.id"), primary_key=True)
+)
+
+class City(Base):
+    __tablename__ = "cities"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String, nullable=False)
+    country = Column(String, nullable=True)
+    latitude = Column(Float, nullable=False)
+    longitude = Column(Float, nullable=False)
+    created_at = Column(DateTime, server_default=func.now(), nullable=False)
+    updated_at = Column(DateTime, onupdate=func.now(), server_default=func.now(), nullable=False)
+    user_visitors = relationship("User", secondary=user_cities, back_populates="visited_cities")
+
+class User(Base):
+    __tablename__ = "users"
+    id = Column(Integer, primary_key=True, index=True)
+    name = Column(String, nullable=False)
+    created_at = Column(DateTime, server_default=func.now(), nullable=False)
+    updated_at = Column(DateTime, onupdate=func.now(), server_default=func.now(), nullable=False)
+    posts = relationship("Post", back_populates="owner")
+    visited_cities = relationship("City", secondary=user_cities, back_populates="user_visitors")
+
+class Post(Base):
+    __tablename__ = "posts"
+    id = Column(Integer, primary_key=True, index=True)
+    title = Column(String, index=True)
+    content = Column(String)
+    owner_id = Column(Integer, ForeignKey("users.id"))
+    created_at = Column(DateTime, server_default=func.now(), nullable=False)
+    updated_at = Column(DateTime, onupdate=func.now(), server_default=func.now(), nullable=False)
+    owner = relationship("User", back_populates="posts")
